+++ frontend/packages/ux-editor/src/components/Properties/PropertiesHeader/HeaderMainConfig.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import { StudioHeading } from '@studio/components';
+import { RequiredIndicator } from '../../RequiredIndicator';
+import classes from './ComponentMainConfig.module.css';
+import { useTranslation } from 'react-i18next';
+
+type HeaderMainConfigProps = {
+  children?: React.ReactNode;
+};
+
+export const HeaderMainConfig = ({ children }: Partial<HeaderMainConfigProps>): JSX.Element => {
+  const { t } = useTranslation();
+
+  return (
+    <div className={classes.componentMainConfig}>
+      <StudioHeading size='2xs'>
+        {t('ux_editor.component_properties.main_configuration')}
+        <RequiredIndicator />
+      </StudioHeading>
+      {children}
+    </div>
+  );
+};
